send notification icons from main process to renderer process
Only the main process can access local resources, like the icon
images. We read the icon images from the main process, encode them
as base64 strings, and send them to the renderer process.

For some reason electron used to be able to read the icon files from
the renderer process, but not anymore. It looks like it was an electron
bug that got fixed at some point.