ignore empty messageTable
On Feb 2020 we noticed that we weren't getting any more new mail
notifications.
This was caused by an empty message table that appeared before and
sometimes after the actual message table. The query $('div.Cp table.F')
was returning the 1st empty table.

A quick workaround is to look for all 'div.Cp table.F' tables and use
the 1st one that has some rows inside.